remove integer comparison warning

+++ lib/eigensolve_quda.cpp
@@ -297,7 +297,7 @@ namespace quda
   void EigenSolver::computeEvals(const DiracMatrix &mat, std::vector<ColorSpinorField *> &evecs,
                                  std::vector<Complex> &evals, int size)
   {
-    if (size > evecs.size() || size > evals.size())
+    if (size > (int)evecs.size() || size > (int)evals.size())
       errorQuda("Requesting %d eigenvalues with only storage allocated for %lu", size, evals.size());
     ColorSpinorParam csParam(*evecs[0]);
     std::vector<ColorSpinorField *> temp;
